Add responsive keyboard
