+++ src/components/ChatRoom/ChatRoom.js
@@ -3,11 +3,13 @@ import { Text, View, StyleSheet, KeyboardAvoidingView } from 'react-native';
 import { connect } from 'react-redux';
 import MessageView from '../MessageView/MessageView';
 import MessageInput from '../MessageInput/MessageInput';
+import { Input, Button } from '../common';
+import { addMessage } from '../../actions';
 
 class ChatRoom extends Component {
 	constructor(props) {
 		super(props);
-		this.state = {};
+		this.state = { message: '' };
 		this.connection = new WebSocket(`http://langchat-crosspollination.herokuapp.com/ws/${this.props.roomId}`);
 	}
 
@@ -26,24 +28,36 @@ class ChatRoom extends Component {
 			console.log('error');
 		};
 		this.connection.onclose = () => {
+			// this.props.sendMessage(this.state.message);
 			console.log('disconnected');
-			// this.connect()
 		};
 	};
 
-	sendMessage = message => {
-		console.log(message);
+	sendMessage = () => {
+		const newMessage = { id: Date.now(), username: 'Jarrett', text: this.state.message };
+		this.props.addMessage(newMessage);
+		this.setState({ message: '' });
 	};
 
 	render() {
 		// const messages = this.props.messages.filter(message => message.room_id === this.props.roomId);
 		return (
-			<View style={styles.ChatRoom}>
-				<KeyboardAvoidingView style={styles.keyboard} behavior="padding">
-					<MessageView messages={this.props.messages} />
-					<MessageInput sendMessage={this.sendMessage} />
-				</KeyboardAvoidingView>
-			</View>
+			<KeyboardAvoidingView style={styles.ChatRoom} behavior="padding" keyboardVerticalOffset={40} enabled>
+				<MessageView messages={this.props.messages} />
+				<View style={styles.inputContainer}>
+					<Input
+						style={styles.messageInput}
+						placeholder="Enter your message here..."
+						value={this.state.message}
+						onChangeText={message => this.setState({ message })}
+					/>
+					<View>
+						<Button onPress={this.sendMessage} style={styles.sendButton}>
+							S
+						</Button>
+					</View>
+				</View>
+			</KeyboardAvoidingView>
 		);
 	}
 }
@@ -53,11 +67,17 @@ const styles = StyleSheet.create({
 		alignContent: 'flex-end',
 		flex: 1
 	},
-	keyboard: {
-		position: 'absolute',
-		bottom: 0,
-		left: 0,
-		right: 0
+	inputContainer: {
+		justifyContent: 'space-between',
+		height: 40
+	},
+	messageInput: {
+		borderColor: 'blue',
+		borderWidth: 1
+	},
+	sendButton: {
+		width: 40,
+		height: 40
 	}
 });
 
@@ -65,4 +85,8 @@ export const mapStateToProps = state => ({
 	messages: state.messages
 });
 
-export default connect(mapStateToProps)(ChatRoom);
+export const mapDispatchToProps = dispatch => ({
+	addMessage: message => dispatch(addMessage(message))
+});
+
+export default connect(mapStateToProps, mapDispatchToProps)(ChatRoom);
