+++ src/components/MessageView/MessageView.js
@@ -12,8 +12,7 @@ const styles = StyleSheet.create({
 	MessageView: {
 		borderColor: 'red',
 		borderWidth: 1,
-    flex: 1,
-    height: 200
+		flex: 4
 	}
 });
